requirements.txt
torch
torchvision
numpy
opencv-python-headless  # using headless version for server deployment
Pillow  # for PIL
openai  # for OpenAI API
gradio
pyyaml  # for yaml
matplotlib
IPython
